feat: add ownership verification

+++ controllers/activity.controllers.js
@@ -8,12 +8,15 @@ router.get('/:tripId/activity/new', isSignedIn, async(req,res)=>{
     try{
         const currentTrip = req.params.tripId
         const foundTrip = await Trip.findById(currentTrip);
+        if(!foundTrip.owner.equals(req.session.user._id)){
+            return res.send('Authorization Failed!')
+        }
         const foundDestinations = await Destination.find({trip:currentTrip});
       
-        console.log(foundTrip)
         res.render('activity/new-activity.ejs', {
             currentTrip: currentTrip,
             destinations: foundDestinations,
+            trips: foundTrip
         })
     }catch(err){
         console.log('ERROR:', err)
@@ -22,6 +25,12 @@ router.get('/:tripId/activity/new', isSignedIn, async(req,res)=>{
 
 router.post('/:tripId/activity/new', isSignedIn, async (req, res)=>{
     try{
+
+        const currentTrip = req.params.tripId
+        const foundTrip = await Trip.findById(currentTrip);
+        if(!foundTrip.owner.equals(req.session.user._id)){
+            return res.send('Authorization Failed!')
+        }
         const dateTime = new Date(`${req.body.dateOfActivity}T${req.body.timeOfActivity}`) //form gives me date and time, so i merged them into one date and stored them as dateTime
         const newActivity = await Activity.create({
             name: req.body.name,
@@ -42,7 +51,12 @@ router.get('/:tripId/activity', isSignedIn, async (req, res)=>{
     try{
 
         const currentTrip = req.params.tripId
-        const foundAllActivities = await Activity.find({trip:req.params.tripId, })
+        const foundTrip = await Trip.findById(currentTrip);
+        console.log(foundTrip)
+        if(!foundTrip.owner.equals(req.session.user._id)){
+            return res.send('Authorization Failed!')
+        }
+        const foundAllActivities = await Activity.find({trip:req.params.tripId })
         res.render('activity/all-activities.ejs', {
             activities: foundAllActivities,
             currentTrip: currentTrip
@@ -55,6 +69,10 @@ router.get('/:tripId/activity', isSignedIn, async (req, res)=>{
 router.get('/:tripId/activity/:activityId', isSignedIn, async (req, res)=>{
     try{
         const currentTrip = req.params.tripId
+        const foundTrip = await Trip.findById(currentTrip);
+        if(!foundTrip.owner.equals(req.session.user._id)){
+            return res.send('Authorization Failed!')
+        }
         const foundOneActivity = await Activity.findById(req.params.activityId).populate('trip').populate('destination')
         console.log(foundOneActivity)
         res.render('activity/activity-details.ejs', {
@@ -71,15 +89,19 @@ router.get('/:tripId/activity/:activityId', isSignedIn, async (req, res)=>{
 
 router.get('/:tripId/activity/:activityId/edit', isSignedIn, async (req, res)=>{
     try{
-        const foundTrips = await Trip.find()
         const currentTrip = req.params.tripId
+        const tripOwner = await Trip.findById((currentTrip))
+        const foundTrip = await Trip.find({owner: req.session.user._id});
+        if(!tripOwner.owner.equals(req.session.user._id)){
+            return res.send('Authorization Failed!')
+        }
         const foundDestinations = await Destination.find({trip: currentTrip})
         const activityTypes = ['Sightseeing', 'Adventure', 'Cultural', 'Relaxation', 'Educational', 'Entertainment', 'Dining']
         const foundOneActivity = await Activity.findById(req.params.activityId).populate('trip').populate('destination')
         res.render('activity/edit-activity.ejs', {
             activity: foundOneActivity,
             currentTrip: currentTrip,
-            trips: foundTrips,
+            trips: foundTrip,
             destinations: foundDestinations,
             types: activityTypes
         })
@@ -89,13 +111,20 @@ router.get('/:tripId/activity/:activityId/edit', isSignedIn, async (req, res)=>{
 })
 
 router.put('/:tripId/activity/:activityId', isSignedIn, async (req, res)=>{
+        const currentTrip = req.params.tripId
+        const tripOwner = await Trip.findById((currentTrip))
+        const foundTrip = await Trip.find({owner: req.session.user._id});
+        if(!tripOwner.owner.equals(req.session.user._id)){
+            return res.send('Authorization Failed!')
+        }
+   
     const dateTime = new Date(`${req.body.dateOfActivity}T${req.body.timeOfActivity}`) //form gives me date and time, so i merged them into one date and stored them as dateTime
     const updatedActivity = await Activity.findByIdAndUpdate(req.params.activityId, {
         name: req.body.name,
         type: req.body.type,
         timeOfActivity: dateTime,
         cost: req.body.cost,
-        trip: req.params.trip,
+        trip: req.body.trip,
         destination: req.body.destination,
     })
 
@@ -104,6 +133,11 @@ router.put('/:tripId/activity/:activityId', isSignedIn, async (req, res)=>{
 
 router.delete('/:tripId/activity/:activityId', isSignedIn, async (req, res)=>{
     try{
+        currentTrip = req.params.tripId
+        const foundTrip = await Trip.findById(currentTrip);
+        if(!foundTrip.owner.equals(req.session.user._id)){
+            return res.send('Authorization Failed!')
+        }
         const deleteActivity = await Activity.findByIdAndDelete(req.params.activityId)
         res.redirect(`/trip/${req.params.tripId}/activity`)
     }catch(err){
