+++ controllers/destination.controllers.js
@@ -6,7 +6,7 @@ const isSignedIn = require('../middleware/is-signed-in')
 router.get('/new', isSignedIn, async (req, res)=>{
     try{
         const currentTrip = req.params.tripId
-        const foundAllTrips = await Trip.find()
+        const foundAllTrips = await Trip.find({owner: req.session.user._id})
         res.render('destination/create-destination.ejs', {
             trips: foundAllTrips,
             currentTrip: currentTrip
@@ -19,6 +19,11 @@ router.get('/new', isSignedIn, async (req, res)=>{
 
 router.post('/new', isSignedIn, async (req,res)=>{
     try{
+        const foundOneTrip = await Trip.findById(req.body.trip)
+        if(!foundOneTrip.owner.equals(req.session.user._id)){
+             return res.send('Authorization Failed!')
+        }
+
         const newDestination = await Destination.create({
             country: req.body.country,
             city: req.body.city,
@@ -34,6 +39,7 @@ router.post('/new', isSignedIn, async (req,res)=>{
 
 router.get('/', isSignedIn, async (req, res)=>{
     try{
+
         const currentTrip = req.params.tripId
         const foundAllDestinations = await Destination.find({owner: req.session.user._id}).populate('trip')
         console.log(foundAllDestinations)
@@ -46,6 +52,10 @@ router.get('/', isSignedIn, async (req, res)=>{
 router.get('/:destinationId/edit', isSignedIn, async(req, res)=>{
     try{
         const foundOneDestination = await Destination.findById(req.params.destinationId).populate('trip')
+        if(!foundOneDestination.owner.equals(req.session.user._id)){
+            return res.send('Authorization Failed')
+        }
+
         res.render('destination/destination-edit.ejs', {destination: foundOneDestination})
     }catch(err){
         console.log('ERROR:', err)
@@ -54,6 +64,11 @@ router.get('/:destinationId/edit', isSignedIn, async(req, res)=>{
 router.put('/:destinationId', isSignedIn, async(req, res)=>{
     
         try{
+        const foundOneDestination = await Destination.findById(req.params.destinationId).populate('trip')
+        if(!foundOneDestination.owner.equals(req.session.user._id)){
+            return res.send('Authorization Failed')
+        }
+
         const updatedDestination = await Destination.findByIdAndUpdate(req.params.destinationId, {
         country: req.body.country,
         city: req.body.city
@@ -67,6 +82,10 @@ router.put('/:destinationId', isSignedIn, async(req, res)=>{
 
 router.delete('/:destinationId', isSignedIn, async (req, res)=>{
     try{
+        const foundOneDestination = await Destination.findById(req.params.destinationId).populate('trip')
+        if(!foundOneDestination.owner.equals(req.session.user._id)){
+            return res.send('Authorization Failed!')
+        }
         const deleteDestination = await Destination.findByIdAndDelete(req.params.destinationId)
         res.redirect('/destination')
     }catch(err){
